requirements.txt
openai>=1.43.0
python-dotenv>=1.0.1
pydantic>=2.7.0
rich>=13.7.1
httpx>=0.27.0
